Increase civicrm_queue_item.data field size
civicrm_queue_item.data field holds information about the task in the
job queue that is going to be excuted. But if we are using queues to
import large data (as we might do in an extension such as
nz.co.fuzion.csvimport) then the current field type might not be able to
handle the data size. So here I am changing it from TEXT to LONGTEXT to
accomdate for such cases.