(REF) CRM_Core_DAO - Simplify getFieldSpec()
Before: This function is complicated to read.  It is also a bit inefficient...
it builds a single-use index of all fields and then forgets it.  That's an O(n)
scan plus O(n) memory allocation and then a lookup on that list.

After: The function is simpler.  It may still do an O(n) scan, but it will stop
early on match, and it won't allocate O(n) memory.