Add deprected warning to apparently unused functions
It looks like one check function is only ever called by the other,
and the other isn't called from anywhere.

What these functions are supposed to do is unclear.
The code comment appears to be wrong; the variable $str
is not a permission string but is inserted into a query
as civicrm_acl.object_table.