[Ref] remove redundant call to clearGroupContactCache
This code does a clear on the cache and then calls the api which will eventually
call load(). Load ALSO clears the cache so this seems redundant.

However, there are a couple of places where add() is called from where it might
serve to force the rebuild due to changes to group make up. In general we have switched
to invalidating and letting the rebuild-on-demand do it's thing to allow
things like crons or lack of need for the group to reduce to browser demand.
So, I think that by switching to invalidating rather than rebuilding
we can remove this potentially locky line from GroupContactCache::add()