[REF] Provide standardised methods to get recurring contribution id & object
This adds methods to get the recurring contribution id & object that work from the paypal
express and from the paypal pro flows. In the paypal express flow it is
retrieved from a known value (and this flow is only ever hit for recurrings
as evidenced by the existing getsingle) and then set on the class. For paypal pro
it is optional and will be in the cheap-to-access POST array if relevant