Fix report buildContactCache to refresh parent groups
This function turns out to have been triggering a refresh / cache inclusion of
smart groups but not parent groups.

The fix also causes it to call load() on each group directly.
Previously it was calling check which calls loadAll
which calls add() which calls the v3 contact api
which calls the query object, which calls load()
on each group