(REF) Extensions - Remove dead code
To see that this code is unused, simply grep the entire source-tree for `upgradable`
and `upgradeVersion`. There is nothing that would plausible set these fields.

I think this is left-over from an older UX. When there's an upgrade available in this UX, it shows
an alert box under the "Description".