(dev/release#17) Extension UI - Show developmental icon for alpha/beta-stage extensions
Extensions are published with two fields, `<version>` (e.g.  `1.1`) and `<develStage>` (e.g.  `alpha`).
Confusingly, the two fields sometimes convey redundant information - but not always. A few cases:

* For example, sometimes a `<version>` as `1.0-alpha1`, `1.2.beta2`, or `2.3dev`. This would make sense if the
developer does formal tagging/releasing for developmental versions. (In this case, you don't really need the
icon - because the version-number tells you.)

* Other times, you might have a `<version>` which simply says `1.0` or `2.0` -- and then supplemental
information where the `<develStage>` says `alpha` or `beta`.  This could make sense if the developmental
version is being continuously updated without formal tags/releases.

* Under dev/release#17, we have another case -- where the core-extensions have version#s which match the
core version# (because they are released together), but the devel-stage of the extension is only
alpha/beta (because the extension is still evolving/optional/not-fully-supported).

Before: The "Development Stage" is entirely obscure.

After: There's an icon in the table to signal the development stage.