Add call to validateAllContributions & fix getTotalAmount
I've updated 2 tests for readability and extended one to call
validateAllContributions. The call failed and it turned out to
be the check was wrong. Fixing that caused a test to fail.
I have fixed the test by fixing getTotalAmount on the internal
helper BAO_Order class to include the tax_amount for each line
in the total.

I think this would most affect tests - but it might affect the
Membership_Form due to changes made recently to that
form so I'm on the fence about going for the rc with this
one