[REF] replace isFirst parameter
Here we see that the 2 paths to 'recur' were calculating isFirst slightly differently.

In one case (the dominant one) we were looking to see if it was completed whereas
in the lesser path (paypal express) we were looking to see if it was not pending.

That leaves
'cancelled' & 'failed' in a limbo. I think logically we can't complete either of those
so the dominant definition is better