For Symfony-style event listeners, encourage use of `void` return
When using Symfony-style listeners, all inputs and outputs for the event go
through the event object.

Note that `hook_*()` notation does allow return values, For these functions,
it is *normal* to return void, but some existing hooks rely on returning
array-data.  It could be misleading if we made it appear that all `hook_*()`
examples have to return void.