Extract code that populates temp table for an individual group
This is part of breaking the individual group actions (this might
actually be the only one that needs to be done on an individual group)
into single actions. Then the ones that could be on one
or more groups can all be consistent.

Note this is only called from one place and is protected (& covered by tests)
so there is no practical change