APIv4 - add is_current as a pseudo (calculated) field
This adds the first calculated field to APIv4. It works by injecting a SQL function and aliasing it
with the name of the fake field. This approach allows it to work as part of the WHERE clause
as well as the SELECT clause, even across joins.

Marking the old "is_current" api param deprecated because the field works better.