Move determination of priceSetID to the internal order class
This moves some functionality from the form to where we can start to use if from the order api.

I also added more 'internal' commenting. This is implicitly the case but
I think it would be tempting to use this class and the comments at the top
of the class saying not to could be missed