dev/event#58 Ensure custom data is not saved to template's row
This ensures than when we are copying (implicity a create-new-action) we do not
update an existing row in the custom data table by explicity unsetting the
id in the copy function. This has been bouncey behaviour because
the formatting for form purposes relies on form type formatting and yet
for save purposes it's not really fit for purpose. Although a bit
hacky this approach should be robust & avoids a re-write on the form