Throw exception when baoName is not found
It the BAO name is not found at this point we currently get a fatal error. This makes it more
recoverable. The specific case I'm hitting at the moment relates to monolog - the
caching of the symfony container means it persists between builds and is
accessed before the extension is installed. However the entity is not
yet registered. Historically I have hit this fatal in other
scenarios but I can't recall the details.