dev/core#2650 Add support for contribution_status_id to the processor
Per https://lab.civicrm.org/dev/core/-/issues/2650 the goal is to get the
token processor (for each entity, in this case contributions) to be processing fields
in the same way as the legacy processor does and for this to be thoroughly tested.

This allows us to expose the token processor classes to message template processing
and use them interchangeably, with an eventual goal of migrating over.

It turns out that althought 'Contribution Status ID' is exposed in the UI
for scheduled reminders it did not work. This left us open to simply add the
token contribution.contribution_status_id and add tests.

Per 2650 the handling of pseudoconstants is inconsistent but our preferred goal
is that they would follow apiv4 style syntax and that the actual field name
would hold the actual field value.

Although the token processor exposes a value 'status' - it seems this is likely never
actually exposed to the user and is probably unused.

We can come back to how to deprecate & remove