Swap displayName smarty for {contact.display_name token}
This swaps over the smartyToken for the token token on the template added on new installs, and adds a test.

Separately I will look at an upgrade script and other templates

To test make sure you enable invoicing & then generate an invoice from search results.
use the demo site due to aforementioned lack of upgrade and check the display name is still on
the left hand side