Activity tasks - don't add invalid tasks
Instead of 'handling the task code building invalid tasks - #20945' - this proposes to only attempt
to add a page if it is for a valid task. This keeps the grouchiness in the generic controller
to warn us of other mistakes.

This code is pretty fugly! However, it really is just called internally within this class.