Fix order api to set price field value for default price set
This makes it such that when using the default price set for contributions it not only
fills the default price field id. I originally had this in the code but I pulled it out
because the line item test failed. However, I'm seeing cases where not
having a price_field_value_id fails to add tax and I think the issue with the line item
test is that it is, incorrecly, trying to add 2 line items on the price field in
the default price set - which is not actually the expectation on the default price set

I do, separately, wonder if we need the unique id for price_field_id + contribution_id
+ price_field_value_id - but out of scope on this