Deprecate db_token back to the only function that calls it
I realised that in the whole git universe only this function calls db_token. The code in
db_token partially duplicates this code and this code is doing work to 'avoid' the stuff db_token
does. I also think the logic of how to direct db token to format the fields belongs in this
function, the signature of db_token is not quite right and the pseudoconstant stuff is wrong.

All in all I think melding this back into it's only calling function and then coming up
with the right shared function makes more sense