Reconcile and upgrade tokens for payment instrument, financial type
This switches the last 2 'alias' tokens from Contribute_Tokens to the preferred format,
adds tests & upgrade and switches the advertised tokens. I tested & the token
removed from the sent letter collection still works

Next step is to analyse the remaining differences between the 2 sets of tokens.
My suspicion is that we have enough metadata that we use in search kit that we
don't need a hard-coded white list of tokens in 2 different places & can just
expose 'all the fields' that are not acl or otherwise limited