APIv4 - Throw exception instead of munging illegal join aliases
The problem with "fixing" an illegal join alias is that it's then mysterious
what the correct alias will be, leading to bugs when the incorrect alias
gets used throughout the rest of the api params.

Throwing an exception seems like a better way to ensure developers
are alerted to the error.