(REF) ActionSchedule - Pass through SQL fields named "tokenContext_*"
This will help us to consolidate the prefetching logic in `CRM_*_Tokens`.

Currently, `CRM_*_Token::alterActionScheduleQuery()` updates the query to select all fields for the entity.
This is regardless of how many fields there are, whether they are needed, etc. This is also prone to naming conflicts.

With this patch, `CRM_*_Token::alterActionScheduleQuery()` only needs to select one field (`tokenContext_fooId`). This
will then propagate to the `TokenProcessor` which can do its own optimized data-loading.