ActionSchedule - Set locale during mail hooks / delivery / etc
From a core POV, this should be unnecessary.  And I'm not sure that it's a
good idea for contrib to assume that the locale has been set to the email
receipient.  (It probably isn't done that way in some contexts.)

Never-the-less, this should provide greater continuity with the pre-existing
behavior.  If anyone, say, uses `Hook::alterMail()` to append a translated
footer, then it should continue to work in the same way as before.