ActionSchedule - Fill TokenProcessor will real batches
Background: `ActionSchedule::sendMailings()` fetches a batch of pending reminders (per
some specific schedule/rule). Then it composes and sends a message for each.

Before: For each item in the batch, it makes a new `TokenProcessor`.  This
means that the `TokenProcessor` only sees one pending reminder. Therefore, the
`TokenProcessor` cannot meaningfully fetch batched data.

After: It creates one `TokenProcessor` and adds rows for each pending
reminder.  This means that `TokenProcessor` can fetch batched data.# On
branch master-action-batch

Comments: This part of a longer plot to simplify the `CRM_*_Tokens`.
Currently, `CRM_*_Tokens` implements duplicate prefetch mechanisms -- the
general-purpose `prefetch()` and the special-purpose
`alterActionScheduleQuery()`.  This patch makes it possible to use
`prefetch()` for all the real work of prefetching (and phasing-out
`alterActionScheduleQuery()`).