dev/core#2749 remove exception when no payment processor configuredhen no payment processor configured
This is an attempt to fix a regression https://lab.civicrm.org/dev/core/-/issues/2749
when there are legitimately no payment processors assigned to a page.

I feel like the configuration screens are adequate to enforce this & having an exception like
this in the assign function is the wrong place.

The original issue is that the is_monetary box was not checked & hence the
processor wasn't assigned. This felt like asking users to understand that
a payment method when money options are configured