[Ref] Copy emailcommon function back to email trait
This is part of ending the email common trait. It is currently 'shared'
with a couple of other places - but most of the functionality is
not that shared & could be easily cleaned up when separated out.

I want to create a trait for PdfTrait like the email trait so
getting the email trait better separated for that